requirements.txt
ansible
cookiecutter==1.0
cryptography
PyYAML
